Remove old routing/group communication library.
As discussed briefly on the list (22.{02,12}.2011), the library was
unused since early this year, and has been deprecated because it:
 * is unable to interface to flounder/THC
 * assumes that a set of core IDs can be expressed in a single-word bitmap
 * doesn't work on machines with large (> 32?) numbers of cores (but
   this is probably just a bug?)

Since we have the new multihop code in the tree, and seem to have
decided that this is an architectural dead-end, I'm removing this code
in the interest of least confusion as we start building functionality
on the new multihop code.

The one major side-effect of this change is that it is no longer
possible to compile the non-NULL remote cap management database
implementations (rcap_db_central, rcap_db_twopc). The code remains in
the tree for future resurrection, however.