stop fixing the start address of the data section (for user programs) on x86_64
This is somewhat experimental. I don't remember why we did this in the
first place (it was when we still used CMake), but the current system
seems to cope fine without it. It's necessary, because with our
current huge text sections (all those static libraries), it's quite
easy for text and data to collide.