simplify out-of-band cap transfer to avoid asynchronously reporting errors back to the sender
As discussed on the mailing list (10.01.2012 "Proposed change to
out-of-band cap transfer"), this change changes the semantics so that
an error in delivery of the cap is reported to the receiver (rather
than asynchronously to the sender) along with a NULL capref. This
simplifies the implementation within the monitor, and results in saner
error handling on both sides.