newlib/x86_32: change newlib's uintptr_t to unsigned long on x86_32
On x86_32, compiling with newlib gives the following error (using clang
because it has better messages):

 ../lib/pci/pci_client.c:175:46: error: incompatible pointer types
   passing 'errval_t *' (aka 'unsigned int *')
   to parameter of type pci_errval_t *' (aka 'unsigned long *')
   [-Werror,-Wincompatible-pointer-types]
     err = pci_client->vtbl.reset(pci_client, &msgerr);

Note that errval_t is uintptr_t (fugu), while pci_errval_t is uint32_t
(flounder).

In freec that's OK:
 include/arch/x86_32/arch/stdint.h: typedef unsigned int uint32_t;
 include/freec/stdint.h: typedef uint32_t uintptr_t;

In newlib on the other hand:
 lib/newlib/newlib/libc/include/stdint.h: typedef unsigned long uint32_t;
 lib/newlib/newlib/libc/include/stdint.h: typedef unsigned __PTRDIFF_TYPE__ uintptr_t;
 On x86_32, __PTRDIFF_TYPE__ is set by the compiler to int...

I think that the clean solution would be to change flounder errvals to uintptr.
However, doing that (if/arch/x86_32.if) makes flounder complain with:

 flounder: cannot compile this interface for UMP;
           it uses intptr/uintptr which are non-portable

So instead, I changed inttypes.h and stdint.h so that uintptr_t is defined as
unsigned long.