thread: fix condition for stack overflow warning.
thread_check_stack_bounds() has a very odd condition.  It will always return
true because bottom <= top.  This results in a warning regardless of whether
the stack pointer is valid or not, for instance:
    Error: stack bounds exceeded: sp = 0x880b00 but [bottom, top] = [0x873000, 0x883000]

We cannot simply patch the condition to instead read:
    return sp > (lvaddr_t)thread->stack &&
           sp <= (lvaddr_t)thread->stack_top;

because that results in failures during bootup:
    >> INFO: stack bounds exceeded: sp = 0x5c1a50 but [bottom, top] = [0x592000, 0x5a2000]
    Dispatcher warning in kernel 0: user trap #13 WHILE DISABLED in 'acpi': IP 477be8, error 0
    skb.0: waiting for: all_spawnds_up
    >> INFO: stack bounds exceeded: sp = 0x631a50 but [bottom, top] = [0x602000, 0x612000]
    Dispatcher warning in kernel 0: user trap #13 WHILE DISABLED in 'kaluga': IP 4a4368, error 0