Octopus: Add ability to set and get records by id capability.
The intent with this patch is that only someone possesing a certain id cap
can retrieve records that were stored with this id cap.

To this end, two new functions were added to the client interface of octopus:

* oct_set_with_idcap(struct capref idcap, const char *attributes, ...)
  Sets a record using the id capability as the key/name of the record.

* oct_get_with_idcap(char **data, struct capref idcap)
  Retrieves a record using the id capability as the key/name.

Note that octopus and the SKB do not support dedicated namespaces atm.
Therefore a record saved with oct_set_with_idcap() could be retrieved by
oct_get().

Moreover, format octopus.if to 80 columns.