New terminal API consisting of libterm_client and libterm_server.
The library libterm_client is responsible for connection setup to a terminal
server as well as character processing. It featues an blocking as well as a
non-blocking API.

The library libterm_server facilitates the development of device drivers that
act as terminal servers.

The API is described in Raphael Fuchs' Bachelor thesis: "A session control
interface for a Multikernel".