MMCHS driver for OMAP4 platform.
Simple driver that can read and write blocks one-at-the-time
on a SDHC card on the pandaboard. We currently don't support SD or MMC cards.
The code was written to for the AOS course at ETH.

Also contains an implementation for the ata_rw28 interface to work with the VFS
FAT implementation.