armv7: Unified global interrupt controller code.
The initialization code for the pl130 global interrupt controller was
duplicated for the arm_gem5 and omap44xx ARMv7 platforms.  This led to bugs
being fixed in one version but not the other. I've now moved all the
controller-specific code to kernel/arch/armv7/gic.c and only left the mapping
and initializing of the GIC registers in the platform-specific code.